requirements.txt
django==2.0.1
psycopg2
psycopg2-binary
Pillow
django-dotenv==1.4.1
djangorestframework-jwt
djangorestframework
